Move elements to shared
