+++ app/src/main/java/com/moxielabs/androidarchitectureexamples/shared/models/Fruit.kt
@@ -1,0 +1,7 @@
+package com.moxielabs.androidarchitectureexamples.shared.models
+
+data class Fruit(
+    val name: String,
+    val color: String,
+    val taste: String
+)
